Don't tag dry-run releases (#239)

+++ src/bump-crates.ts
@@ -9,6 +9,7 @@ import { gitEnv } from "./config";
 
 export type Input = {
   version: string;
+  liveRun: boolean;
   branch: string;
   repo: string;
   path?: string;
@@ -20,6 +21,7 @@ export type Input = {
 
 export function setup(): Input {
   const version = core.getInput("version", { required: true });
+  const liveRun = core.getBooleanInput("live-run", { required: true });
   const branch = core.getInput("branch", { required: true });
   const repo = core.getInput("repo", { required: true });
   const path = core.getInput("path");
@@ -30,6 +32,7 @@ export function setup(): Input {
 
   return {
     version,
+    liveRun,
     branch,
     repo,
     path: path === "" ? undefined : path,
@@ -71,8 +74,11 @@ export async function main(input: Input) {
     }
 
     sh(`git push --force ${remote} ${input.branch}`, { cwd: repo });
-    sh(`git tag --force ${input.version} --message v${input.version}`, { cwd: repo, env: gitEnv });
-    sh(`git push --force ${remote} ${input.version}`, { cwd: repo });
+
+    if (input.liveRun) {
+      sh(`git tag --force ${input.version} --message v${input.version}`, { cwd: repo, env: gitEnv });
+      sh(`git push --force ${remote} ${input.version}`, { cwd: repo });
+    }
 
     sh("git log -10", { cwd: repo });
     sh("git show-ref --tags", { cwd: repo });
